OVS-3695

+++ ci/tests/vmachine/vmachine_test.py
@@ -23,35 +23,35 @@ from ci import autotests
 testsToRun = general.get_tests_to_run(autotests.get_test_level())
 
 VPOOL_NAME = test_config.get('vpool', 'vpool_name')
+TEMPLATE_SERVERS = ['http://sso-qpackages-loch.cloudfounders.com/templates/openvstorage', 'http://172.20.3.8/templates/openvstorage']
 
-template_servers = ['http://sso-qpackages-loch.cloudfounders.com/templates/openvstorage', 'http://172.20.3.8/templates/openvstorage']
-template_folder = '/fio_debian/'
+
+template_source_folder = '/fio_debian/'
 template_image = 'debian.qcow2'
-templateDir = '/var/tmp/templates/'
+template_target_folder = '/var/tmp/templates/'
 
 NUMBER_OF_DISKS = 10
 GRID_IP = test_config.get('main', 'grid_ip')
 
 
 def download_template(server_location):
-    general.execute_command('wget -P {0} {1}{2}{3}'.format(templateDir, template_servers[server_location], template_folder, template_image))
-    general.execute_command('chown root {0}{1}'.format(templateDir, template_image))
+    general.execute_command('wget -P {0} {1}{2}{3}'.format(template_target_folder, server_location, template_source_folder, template_image))
+    general.execute_command('chown root {0}{1}'.format(template_target_folder, template_image))
 
 
-def check_loch_axs():
-    if GRID_IP.split('.')[0] == '172' and GRID_IP.split('.')[1] == '20':
-        return 1
+def get_template_location_by_ip(ip):
+    if ip.split('.')[0] == '172' and ip.split('.')[1] == '20':
+        return TEMPLATE_SERVERS[1]
     else:
-        return 0
+        return TEMPLATE_SERVERS[0]
 
 
 def check_template_exists():
-    out, err = general.execute_command('[ -d {0} ] && echo "Dir exists" || echo "Dir does not exists"'.format(templateDir))
+    out, err = general.execute_command('[ -d {0} ] && echo "Dir exists" || echo "Dir does not exists"'.format(template_target_folder))
     if 'not' not in out:
-        general.execute_command('rm -rf {0}'.format(templateDir))
-        general.execute_command('mkdir {0}'.format(templateDir))
-    server_location = check_loch_axs()
-    download_template(server_location)
+        general.execute_command('rm -rf {0}'.format(template_target_folder))
+        general.execute_command('mkdir {0}'.format(template_target_folder))
+    download_template(get_template_location_by_ip(GRID_IP))
 
 
 def setup():
@@ -61,8 +61,10 @@ def setup():
 
 
 def teardown():
-    general.execute_command("rm -rf /mnt/{0}".format(VPOOL_NAME))
-    general.api_remove_vpool(VPOOL_NAME)
+    vpool = VPoolList.get_vpool_by_name(VPOOL_NAME)
+    if vpool:
+        general.execute_command("rm -rf /mnt/{0}".format(VPOOL_NAME))
+        general.api_remove_vpool(VPOOL_NAME)
     vpool_test.teardown()
 
 
@@ -78,4 +80,6 @@ def vms_with_fio_test():
     # @TODO change this to api connection call
     vpool = VPoolList.get_vpool_by_name(VPOOL_NAME)
     for disk_number in range(NUMBER_OF_DISKS):
-        general.execute_command('qemu-img convert -O raw {0}{1} /mnt/{2}/disk-{3}.raw'.format(templateDir, template_image, vpool.name, disk_number))
+        general.execute_command('qemu-img convert -O raw {0}{1} /mnt/{2}/disk-{3}.raw'.format(template_target_folder, template_image, vpool.name, disk_number))
+
+    assert len(vpool.vdisks) == NUMBER_OF_DISKS, "Only {0} out of {1} VDisks have been created".format(len(vpool.vdisks), NUMBER_OF_DISKS)
